refactor : apiresponse 관련 수정 - 기존 컨트롤러, 서비스 수정 #9
